+++ src/main/config/remotes-config.ts
@@ -1,0 +1,6 @@
+import { RemoteInfo } from "../../shared/models/RemoteInfo";
+
+/** configuration of the remotes (stored in file-system) */
+export class RemotesConfig {
+  remoteInfos: RemoteInfo[] = [];
+}